cargo build passes on Linux
